+++ Runtime/Views/LineView.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 #if USE_TMP
 using TMPro;
 #else
-using TextMeshProUGUI = TMPShim;
+using TextMeshProUGUI = Yarn.Unity.TMPShim;
 #endif
 
 namespace Yarn.Unity