+++ Runtime/Views/OptionView.cs
@@ -4,7 +4,7 @@ using UnityEngine.EventSystems;
 #if USE_TMP
 using TMPro;
 #else
-using TextMeshProUGUI = TMPShim;
+using TextMeshProUGUI = Yarn.Unity.TMPShim;
 #endif
 
 namespace Yarn.Unity